cape: display: Optimize e667 power on/off sequence and regulators
According to the e667 datasheet:
1. Nominal VDDIO is 1.8V
2. Nominal VCI is 3.3V
3. A 96 ms delay is needed after sending the sleep-out command (0x11)
4. A delay of 2 VS (vertical sync, i.e., the time needed to render 1 frame)
   is needed after sending the display-on command (0x29)
5. There's no requirement anywhere to send the write-display-brightness
   command (0x51) upon powering off the panel
6. A delay of 2 VS is needed after sending the display-off command (0x28)
7. A delay of 1 VS is needed after sending the sleep-in command (0x10)
8. A 10 ms delay is needed after enabling VCI and asserting the reset pin
   before the reset pin can be released

Configure all of these as such to match the datasheet.

Additionally, according to the vtdr6130 datasheet:
1. The time between enabling VDDIO and VCI should be 1 ms
2. The time between disabling VCI and VDDIO should be 1 ms
3. The reset pin must be released for at least 15 us to trigger a reset
4. A 15 ms delay is needed after triggering a reset before sending commands

Configure these parameters to match the datasheet as well. Since we can
only configure the reset sequence delays with millisecond granularity, use
a 1 ms delay after releasing the reset pin (which only needs to be released
for 15 us).

Furthermore, since the AVDD, ELVDD, and ELVSS regulators are managed in
hardware using the SWIRE protocol, there's no point in having a post-on
delay for these regulators. Remove the 20 ms post-on delay from the IBB
(ELVSS) regulator.

Change-Id: I7b2e9fe3d99d06b44724d886655daf7deb950577
Signed-off-by: Sultan Alsawaf <sultan@osomprivacy.com>